src/django/requirements.txt
django
python-dotenv
requests
uvicorn[standard]
django-cors-headers
djangorestframework
whitenoise
pygments